requirements.txt
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-meta==2.0.0
Pillow==7.0.0
django-custom-cache-page==0.3
django-robots==4.0
django-pwa